Global reworking

+++ commands/init.js
@@ -1,12 +1,19 @@
 import process from "process"
 import crypto from "crypto"
 import fs from "fs"
-
 import input from "#lib/input"
+import colorize from "#lib/colorize"
 
 const current_directory = process.cwd()
 
 export default async options => {
+  console.log(
+    colorize(
+      "Welcome to hims initialization script, if u want to create a default config run script with flag --default",
+      "cyan"
+    )
+  )
+
   const config = {}
 
   if (options.default) {
@@ -17,7 +24,7 @@ export default async options => {
     config.admin_initialisation_password = "hims"
     config.secret_user_key = crypto.randomBytes(32).toString("hex")
     config.secret_admin_key = crypto.randomBytes(32).toString("hex")
-    config.mongo_username = "hims"
+    config.mongo_username = "root"
     config.mongo_password = crypto.randomBytes(12).toString("hex")
     config.mongo_database = "hims"
     config.redis_password = crypto.randomBytes(12).toString("hex")
@@ -29,11 +36,11 @@ export default async options => {
     config.project_name = await input("Project name", "Simple HiMS App")
     config.project_description = await input("Project description", "A simple HiMS app created with the HiMS CLI")
     config.environment = await input("Environment", "development")
-    config.admin_initialisation_username = await input("Admin initialisation username", "hims")
-    config.admin_initialisation_password = await input("Admin initialisation password", "hims")
+    config.admin_initialisation_username = await input("Admin initialisation username")
+    config.admin_initialisation_password = await input("Admin initialisation password")
     config.secret_user_key = crypto.randomBytes(32).toString("hex")
     config.secret_admin_key = crypto.randomBytes(32).toString("hex")
-    config.mongo_username = await input("MongoDB username", "hims")
+    config.mongo_username = await input("MongoDB username", "root")
     config.mongo_password = crypto.randomBytes(12).toString("hex")
     config.mongo_database = await input("MongoDB database", "hims")
     config.redis_password = crypto.randomBytes(12).toString("hex")
@@ -43,10 +50,23 @@ export default async options => {
     config.admin_path = await input("Admin path", "admin")
   }
 
+  const configPath = `${current_directory}/.hims.env`
+
+  if (fs.existsSync(configPath)) {
+    console.log(colorize("Configuration file already exists!", "red"))
+    const overwrite = await input(colorize("Do you want to overwrite the configuration? (y/No): ", "blue"))
+    if (!["yes", "y"].includes(overwrite.trim().toLowerCase())) {
+      console.log("Exiting without changes.")
+      return
+    }
+  }
+
   fs.writeFileSync(
-    `${current_directory}/.hims.env`,
+    configPath,
     Object.entries(config)
       .map(([key, value]) => `${key.toUpperCase()}=${value}`)
       .join("\n")
   )
+
+  console.log("Configuration saved successfully.\nHappy coding!\n")
 }
